requirements.txt
google-cloud-bigquery==3.12.0
google-cloud-storage==2.12.0
